Clean up in the TestOrder/TestEvent world.
* fixed an issue where a workaround for a bug (#677) was eliminated
  due to a misunderstanding
* eliminated dead repository method and service method
* made test from deleted repository method live on as a test of the
  new method that more or less replaces it (with some refactoring)
