+++ backend/src/test/java/gov/cdc/usds/simplereport/db/repository/TestOrderRepositoryTest.java
@@ -52,7 +52,11 @@ public class TestOrderRepositoryTest extends BaseRepositoryTest {
         assertEquals(0, queue.size());
         queue = _repo.fetchQueue(gtown, site);
         assertEquals(1, queue.size());
-        doTest(order, TestResult.NEGATIVE);
+        TestEvent event = _dataFactory.doTest(order, TestResult.NEGATIVE);
+        flush();
+        TestOrder lookuporder = _repo.findByTestEventId(order.getOrganization(), event.getInternalId());
+        assertNotNull(lookuporder);
+        assertEquals(lookuporder.getInternalId(), order.getInternalId());
         assertEquals(0, _repo.fetchQueue(gtown, site).size());
         assertEquals(1, _repo.fetchPastResults(gtown, site).size());
     }
@@ -163,55 +167,5 @@ public class TestOrderRepositoryTest extends BaseRepositoryTest {
         assertEquals("Charles", orders.get(0).getPatient().getFirstName());
         assertEquals("Adam", orders.get(1).getPatient().getFirstName());
         assertEquals("Bradley", orders.get(2).getPatient().getFirstName());
-    }
-
-    @Test
-    public void fetchResults_multipleEntries_sortedLifo() throws InterruptedException {
-        Organization org = _dataFactory.createValidOrg();
-        Person adam = _dataFactory.createMinimalPerson(org, null, "Adam", "A.", "Astaire", "Jr.");
-        Person brad = _dataFactory.createMinimalPerson(org, null, "Bradley", "B.", "Bones", null);
-        Person charlie = _dataFactory.createMinimalPerson(org, null, "Charles", "C.", "Crankypants", "3rd");
-        Facility facility = _dataFactory.createValidFacility(org);
-
-        TestOrder charlieOrder = _repo.save(new TestOrder(charlie, facility));
-        pause();
-        TestOrder adamOrder = _repo.save(new TestOrder(adam, facility));
-        pause();
-        TestOrder bradleyOrder = _repo.save(new TestOrder(brad, facility));
-
-        List<TestOrder> results = _repo.fetchPastResults(org, facility);
-        assertEquals(0, results.size());
-
-        doTest(bradleyOrder, TestResult.NEGATIVE);
-        pause();
-        doTest(charlieOrder, TestResult.POSITIVE);
-        pause();
-        doTest(adamOrder, TestResult.UNDETERMINED);
-
-        results = _repo.fetchPastResults(org, facility);
-        assertEquals(3, results.size());
-        assertEquals("Adam", results.get(0).getPatient().getFirstName());
-        assertEquals("Charles", results.get(1).getPatient().getFirstName());
-        assertEquals("Bradley", results.get(2).getPatient().getFirstName());
-    }
-
-    private void doTest(TestOrder order, TestResult result) {
-        order.setResult(result);
-        TestEvent event = _events.save(new TestEvent(order));
-        order.setTestEventRef(event);
-        order.markComplete();
-        _repo.save(order);
-        flush();
-        TestOrder lookuporder = _repo.findByTestEventId(order.getOrganization(), event.getInternalId());
-        assertNotNull(lookuporder);
-        assertEquals(lookuporder.getInternalId(), order.getInternalId());
-    }
-
-    private static void pause() {
-        try {
-            Thread.sleep(2);
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
     }
 }
