+++ backend/src/test/java/gov/cdc/usds/simplereport/service/TestOrderServiceTest.java
@@ -127,32 +127,11 @@ public class TestOrderServiceTest extends BaseServiceTest<TestOrderService> {
 
         @Test
         @WithSimpleReportStandardUser
-        public void fetchTestResults_standardUser_ok() {
-                Organization org = _organizationService.getCurrentOrganization();
-                Facility facility = _organizationService.getFacilities(org).get(0);
-                Person p = _dataFactory.createFullPerson(org);
-                TestEvent _e = _dataFactory.createTestEvent(p, facility);
-
-                // Count queries with one order
-                hibernateQueryInterceptor.startQueryCount();
-                _service.getTestResults(facility.getInternalId());
-                assertEquals(10, hibernateQueryInterceptor.getQueryCount());
-
-                // Count queries with three order
-                TestEvent _e1 = _dataFactory.createTestEvent(p, facility);
-                TestEvent _e2 = _dataFactory.createTestEvent(p, facility);
-                hibernateQueryInterceptor.startQueryCount();
-                _service.getTestResults(facility.getInternalId());
-                assertEquals(12, hibernateQueryInterceptor.getQueryCount());
-        }
-
-        @Test
-        @WithSimpleReportStandardUser
         public void fetchTestEventsResults_standardUser_ok() {
                 Organization org = _organizationService.getCurrentOrganization();
                 Facility facility = _organizationService.getFacilities(org).get(0);
                 Person p = _dataFactory.createFullPerson(org);
-                TestEvent _e = _dataFactory.createTestEvent(p, facility);
+                _dataFactory.createTestEvent(p, facility);
 
                 // Count queries with one order
                 hibernateQueryInterceptor.startQueryCount();
@@ -160,8 +139,8 @@ public class TestOrderServiceTest extends BaseServiceTest<TestOrderService> {
                 long startQueryCount = hibernateQueryInterceptor.getQueryCount();
 
                 // Count queries with three order
-                TestEvent _e1 = _dataFactory.createTestEvent(p, facility);
-                TestEvent _e2 = _dataFactory.createTestEvent(p, facility);
+                _dataFactory.createTestEvent(p, facility);
+                _dataFactory.createTestEvent(p, facility);
                 hibernateQueryInterceptor.startQueryCount();
                 _service.getTestEventsResults(facility.getInternalId(), new Date(0));
                 long endQueryCount = hibernateQueryInterceptor.getQueryCount();
@@ -203,9 +182,7 @@ public class TestOrderServiceTest extends BaseServiceTest<TestOrderService> {
                 assertEquals(1, events_before.size());
 
                 // verify the original order was updated
-                List<TestOrder> savedOrders = _service.getTestResults(facility.getInternalId());
-                assertEquals(1, savedOrders.size());
-                TestOrder onlySavedOrder = savedOrders.get(0);
+                TestOrder onlySavedOrder = _service.getTestResult(_e.getInternalId()).getTestOrder();
                 assertEquals(reasonMsg, onlySavedOrder.getReasonForCorrection());
                 assertEquals(deleteMarkerEvent.getInternalId().toString(), onlySavedOrder.getTestEventId().toString());
                 assertEquals(TestCorrectionStatus.REMOVED, onlySavedOrder.getCorrectionStatus());
